+++ tests/testthat/test-acceptance.R
@@ -1,0 +1,3 @@
+skip_if_not(covr::in_covr())
+skip_on_cran()
+cucumber::test("../acceptance")